+++ src/utils/Dict.h
@@ -60,7 +60,7 @@ public:
 
     int             Intern(const char *s, bool *alreadyPresent=nullptr);
     size_t          StringsCount() const { return intToStr.Count(); }
-    const char *    GetByIndex(int n) const { return intToStr.At(n); }
+    const char *    GetByIndex(size_t n) const { return intToStr.At(n); }
 
     int             nInternCalls; // so we know how effective interning is
 };
